\name{basic.qc}
\alias{basic.qc}

\title{Create quality control plots.}

\description{
  Create quality control plots, used in the examples of CTL mapping.
}

\usage{
basic.qc(genotypes, phenotypes, map_info)
}

\arguments{
  \item{genotypes}{ Matrix of genotypes. (individuals x markers) }
  \item{phenotypes}{ Matrix of phenotypes. (individuals x phenotypes) }
  \item{map_info}{ Matrix of genetic map information }
}

\details{
  None.
}

\value{
  None. 
}

\author{ 
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}

\examples{
  #TODO
}

\seealso{
  \itemize{
    \item \code{\link{CTLscan}} - Scan for CTL
    \item \code{\link{plot.CTLscan}} - Plot a CTLscan object
  }
}

\references{
  TODO
}

\keyword{print}
