\name{ctl-internal}

%ctl.network.R
\alias{CTLnetwork.addmarker}

%ctl.correlation.R
\alias{correlation}
\alias{chiSQN}
\alias{test.correlation}
\alias{test.chiSQN}

%ctl.power.R
\alias{correlated.phenotype}
\alias{check.marker}
\alias{check.pheno}
\alias{create.marker}
\alias{CTLpowerAnalysis}
\alias{dcor.create}
\alias{getRatio}
\alias{plot.CTLnetwork}
\alias{test.power.test}
\alias{CTLanalyseRatios}
\alias{CTLanalyseEffects}

%ctl.circle.R
\alias{mapinfotomarkerlocs}
\alias{draw.element}
\alias{draw.spline}
\alias{circle.loc}
\alias{nfrom}
\alias{nto}

%ctl.statistics.R
\alias{extrapolateBeyondRange}
\alias{CTLscantoScanone}
\alias{CTLtoLODvector}

%ctl.sd.R
\alias{deltaSD}

%ctl.slope.R
\alias{stdSlopeEstimate}
\alias{stdSlopeIntercept}
\alias{printSignificantSlope}
\alias{stdSlopeError}
\alias{stdSlopeBeta}
\alias{stdSlopeTest}

%ctl.utils.R
\alias{munique}
\alias{check.genotypes}
\alias{getRVM}
\alias{lodscorestoscanone}
\alias{gcLoop}

%ctl.image.R
\alias{internal.image}
\alias{addChromosomeLines}

%ctl.print.R
\alias{getPermuteThresholds}
\alias{print.CTLpermute}

%ctl.plot.R
\alias{chr_length}
\alias{chr_total_length}
\alias{a_loc}
\alias{m_loc}
\alias{plot.CTLscan2}
\alias{plot.CTLscan3}
\alias{plotExpression}
\alias{CTLasLOD}

%helper.functions.R
\alias{remove.diag}
\alias{dcor}
\alias{get.chr.edges}
\alias{up}
\alias{dw}
\alias{redblue}
\alias{whiteblack}
\alias{top.correlated}

\title{Internal CTL functions}

\description{
  Internal CTL functions. These are generally not to be called by the user.
}

\author{Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
