\name{CTLhelper}
\alias{ctl.name}
\alias{ctl.names}
\alias{ctl.ctlmatrix}
\alias{ctl.qtlmatrix}
\alias{ctl.lodmatrix}
\alias{ctl.qtlprofile}
\title{CTLhelper - Helper functions for CTL mapping }
\description{
  Helper functions for Correlated Trait Locus (CTL) mapping
}
\usage{
ctl.names(CTLobject)
ctl.qtlmatrix(CTLobject)

ctl.name(CTLscan)
ctl.ctlmatrix(CTLscan)
ctl.lodmatrix(CTLscan)
ctl.qtlprofile(CTLscan)
}

\arguments{
  \item{CTLobject}{  An object of class \code{"CTLobject"}, as output by \code{\link{CTLscan}}. }
  \item{CTLscan}{ An object of class \code{"CTLscan"}. This is a single element from an \code{"CTLobject"}, as output by \code{\link{CTLscan}}. }
}
\details{
TODO
}
\value{
TODO
}
\references{ 
  TODO
}
\author{ 
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\note{
  TODO
}
\examples{
  #TODO
}

\keyword{methods}
