\name{print.CTLobject}
\alias{print.CTLobject}
\title{Print the results of a CTL genome scan}
\description{
  Print the results of a multiple phenotype CTL genome scan produced by \code{\link{CTLscan}}.
}
\usage{
  \method{print}{CTLobject}(x, \dots)
}
\arguments{
 \item{x}{       An object of class \code{"CTLobject"}, as output by \code{\link{CTLscan}}. }
 \item{\dots}{   Ignored.} 
}
\details{
  None.
}
\value{
  None. 
}
\author{ 
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}
\examples{
  #TODO
}
\seealso{
  \itemize{
    \item \code{\link{CTLscan}} - Scan for CTL
    \item \code{\link{plot.CTLscan}} - Plot a CTLscan object
  }
}
\references{
  TODO
}
\keyword{print}
