\name{print.CTLscan}
\alias{print.CTLscan}

\title{Print the results of a CTL genome scans}

\description{
  Print the results of a \code{\link{CTLscan}}
}

\usage{
\method{print}{CTLscan}(x, \dots)
}

\arguments{
 \item{x}{ An object output by the function \code{\link{CTLscan}}.}
 \item{\dots}{ Ignored.} 
}

\details{
  None.
}

\value{
  None. 
}

\author{ 
  Danny Arends \email{Danny.Arends@gmail.com}\cr
  Maintainer: Danny Arends \email{Danny.Arends@gmail.com} 
}

\examples{
  #TODO
}

\seealso{
  \itemize{
    \item \code{\link{CTLscan}} - Scan for CTL
    \item \code{\link{plot.CTLscan}} - Plot a CTLscan object
  }
}

\references{
  TODO
}

\keyword{print}
