\name{ctl-internal}

%ctl.network.R
\alias{write.nodeAttributeFile}
\alias{KEGG.pathway.parser}
\alias{nodesToGenes}

%zzz.r
\alias{has_d}
\alias{has_snow}
\alias{has_rcurl}
\alias{has_rqtl}
\alias{get_CtlEnv}

%ctl.statistics.r
\alias{CTLpermute.internal}
\alias{CTLpermuteMC}
\alias{CTLscantoScanone}
\alias{CTLscoretoPvalue}
\alias{toLod}
\alias{CTLasLOD}
\alias{CTLtoLODvector}
\alias{CTLtoP}
\alias{extrapolateBeyondRange}
\alias{CTLtoPvalue.internal}

%ctl.utils.R
\alias{check.genotypes}
\alias{getRVM}
\alias{lodscorestoscanone}
\alias{gcLoop}

%ctl.image.R
\alias{internal.image}
\alias{qtlimage}

%qtl.scan.R
\alias{QTLscan}
\alias{create.matrix}
\alias{clean.phenotypes}
\alias{map.fast}
\alias{QTLscan.cross}

%ctl.print.R
\alias{getPermuteThresholds}
\alias{print.CTLpermute}

%correlations.R
\alias{matrix_correlation}

%ctl.plot.R
\alias{plot.CTLscan2}
\alias{plotExpression}


\title{Internal CTL functions}

\description{
  Internal CTL functions. These are generally not to be called by the user.
}

\author{Danny Arends, \email{Danny.Arends@Gmail.com} }

\keyword{internal}
